code/python/requirements.txt
# python/requirements.txt
# LLM / Agent Stack
httpx>=0.27.0
# If you add LangChain later:
# langchain>=0.3.0
# langchain-community>=0.3.0

# gRPC Bridge
ggrpc>=1.6.0
gprotobuf>=4.25.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Observability
prometheus-client>=0.21.0
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0          # AsyncHTTPTransport for LLM client testing
